+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### (unreleased)
 
+* Add finder.Folder method
+
 * Add object.Common.Destroy method
 
 * Add object.ComputeResource.Reconfigure method